Revert "Add initial proper eco bank support."
This reverts commit 9c838ae8
