+++ src/com/palmergames/bukkit/towny/object/Government.java
@@ -1,6 +1,5 @@
 package com.palmergames.bukkit.towny.object;
 
-import com.palmergames.bukkit.towny.TownyMessaging;
 import com.palmergames.bukkit.towny.TownySettings;
 import com.palmergames.bukkit.towny.event.GovernmentTagChangeEvent;
 import com.palmergames.bukkit.towny.exceptions.EconomyException;
@@ -10,7 +9,7 @@ import com.palmergames.bukkit.towny.invites.InviteHandler;
 import com.palmergames.bukkit.towny.invites.exceptions.TooManyInvitesException;
 import com.palmergames.bukkit.towny.object.economy.AccountAuditor;
 import com.palmergames.bukkit.towny.object.economy.BankEconomyHandler;
-import com.palmergames.bukkit.towny.object.economy.Bank;
+import com.palmergames.bukkit.towny.object.economy.BankAccount;
 import com.palmergames.bukkit.towny.object.economy.GovernmentAccountAuditor;
 import com.palmergames.util.StringMgmt;
 import org.bukkit.Bukkit;
@@ -23,7 +22,7 @@ import java.util.List;
 
 public abstract class Government extends TownyObject implements BankEconomyHandler, ResidentList, TownyInviter, SpawnLocation {
 	
-	protected Bank account;
+	protected BankAccount account;
 	protected Location spawn;
 	protected String tag = "";
 	protected String board = null;
@@ -157,11 +156,11 @@ public abstract class Government extends TownyObject implements BankEconomyHandl
 	}
 
 	@Override
-	public Bank getAccount() {
+	public BankAccount getAccount() {
 		if (account == null) {
 			String accountName = StringMgmt.trimMaxLength(getEconomyPrefix() + getName(), 32);
 			World world = getWorld();
-			account = new Bank(accountName, world, getBankCap());
+			account = new BankAccount(accountName, world, getBankCap());
 			account.setAuditor(accountAuditor);
 		}
